requirements.txt
pandas~=2.3.1
tqdm~=4.67.1
dvc[webdav]~=3.62.0
pyyaml~=6.0.2
torch~=2.8.0
brevitas~=0.11.0
numpy<2.0.0
h5py~=3.14.0
einops~=0.8.1
onnx==1.17.0
